moving replyto support out of receiver side distribution and moving it to partition configurion for local endpoint
